requirements.txt
streamlit>=1.22.0
pandas>=1.5.3
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
plotly>=5.13.1
pydantic>=2.0.0
bcrypt>=4.0.1
python-jose>=3.3.0
reportlab>=3.6.12
pillow>=9.5.0
openpyxl>=3.1.2
python-dateutil>=2.8.2
requests>=2.28.2
email-validator>=2.0.0
streamlit-option-menu>=0.3.2
streamlit-calendar>=0.1.0
qrcode>=7.3.1
alembic>=1.12.0  # Para migraciones de base de datos
psycopg2>=2.9.9  # Driver PostgreSQL nativo
SQLAlchemy-Utils>=0.41.1  # Utilidades adicionales para SQLAlchemy
python-jose[cryptography]>=3.3.0  # Para manejo de JWT
passlib>=1.7.4  # Para hash de contraseñas
python-multipart>=0.0.6  # Para manejo de formularios
uvicorn>=0.15.0  # Servidor ASGI
gunicorn>=20.1.0  # Servidor WSGI para producción
